+++ e2e-testing/__tests__/helpers/types.ts
@@ -1,0 +1,9 @@
+export type OrderDetails = {
+  mnemonic: string;
+  timeInForce: number;
+  orderFlags: number;
+  side: number;
+  clobPairId: number;
+  quantums: number;
+  subticks: number;
+};
